# openapi/protocol-v1.yaml
openapi: 3.1.0
info:
  title: unsea Protocol API v1
  version: 1.0.0-draft
  description: |
    新一代 WhatsApp 协议层接口（基于 Baileys 7.x）。
    本规范覆盖《Baileys 协议层接口封装与 2000 账号承载方案》§ 3 + 附录 A。

    ## 设计原则

    - 协议层只负责 ws / Noise / creds / 重连，业务编排在 Java 业务层
    - 所有状态查询都带 evidence（时间戳 + 来源）
    - 所有重连相关都对应明确的语义码（PROXY_FAILED / RATE_LIMITED / NEED_REAUTH ...）
    - 接口契约稳定，protobuf 内部码不外泄

    ---

    ## 业务场景导航

    ### 🔑 账号登录 / 授权（3 种方式 + 4 种导入）

    | 业务场景 | 接口 | 说明 |
    |---|---|---|
    | 首次绑定新号（pairing code） | `POST /v1/auth/pairing-code` | 用户在手机 WA "已登录设备" 里输入 8 位码；90s 不绑会自动释放 |
    | 首次绑定新号（扫码） | `POST /v1/auth/qrcode` | 用户用手机 WA 扫返回的二维码；180s 不绑会自动释放 |
    | 标准 Baileys JSON 导入（推荐） | `POST /v1/accounts/import/baileys-json` | 完整 creds + keys，导入即可上线 |
    | 全参数登录（30+ 字段） | `POST /v1/accounts/import/params` | 移植自 Apifox `paramsLogin` |
    | 六段参数登录 | `POST /v1/accounts/import/six` | 移植自 Apifox `sixLogin`，含 deviceIdentityKey |
    | 旧 JSON 登录 | `POST /v1/accounts/import/legacy-json` | 移植自 Apifox `jsonLogin`，base64 |
    | 批量导入（最多 500） | `POST /v1/accounts/import/batch` | 含冷启动节流，避免一次性灌爆 worker |
    | 短信注册新号 | （不支持） | WA 业务不开放此能力 |

    **关键概念**：换 IP 不需要重新授权（详见 § 4.3）。一次 pairing 后 creds 永久持有，重连用本地 creds 即可。

    ### 🚪 账号生命周期

    | 业务场景 | 接口 |
    |---|---|
    | 已绑账号上线（用已存 creds） | `POST /v1/accounts/{accountId}/online` |
    | 主动下线（保留 creds，可重新 online） | `POST /v1/accounts/{accountId}/offline` |
    | 退出并移除设备 | `POST /v1/accounts/{accountId}/logout` |

    ### 🏷️ 账号状态查询

...

...

...

...

...

...

...

...

...

...

...
servers:
...
paths:
...
  /v1/auth/pairing-code:
    post:
      tags: [auth]
      summary: 首次绑定取 pairing code
      description: |
        异步接口。返回 code 通过事件 `pairing.code_generated` 推送。
        Baileys: `sock.requestPairingCode(phone, customPairingCode)`。
        `customPairingCode` 不传时，协议层会优先使用环境变量 `PAIRING_DEFAULT_CODE`；也未配置时由 Baileys 随机生成。
      operationId: requestPairingCode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, proxy]
              properties:
                phone:
                  type: string
                  pattern: '^[0-9]{10,15}$'
                  example: "8617600627277"
                clientRefId:
                  type: string
                  description: 业务侧关联 ID（兼容 Apifox qrId）
                customPairingCode:
                  type: string
                  pattern: '^[A-Za-z0-9]{8}$'
                  description: 自定义 8 位 code（可选）。不传则使用服务端 `PAIRING_DEFAULT_CODE`，再没有才随机。
                  example: "12345678"
                deviceProfile:
                  $ref: '#/components/schemas/DeviceProfile'
                browserDisplay:
                  $ref: '#/components/schemas/BrowserDisplay'
                proxy:
                  $ref: '#/components/schemas/ProxyBinding'
      responses:
        '202':
          description: 已发起 pairing 流程
          content:
            application/json:
              schema:
                type: object
                properties:
                  accountId: { type: string }
                  pairingId: { type: string }
                  expiresAt: { type: string, format: date-time }
  /v1/auth/qrcode:
    post:
      tags: [auth]
      summary: 首次绑定取二维码
      operationId: requestQrCode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [proxy]
              properties:
                clientRefId: { type: string }
                deviceProfile:
                  $ref: '#/components/schemas/DeviceProfile'
                browserDisplay:
                  $ref: '#/components/schemas/BrowserDisplay'
                proxy:
                  $ref: '#/components/schemas/ProxyBinding'
      responses:
        '202':
          description: 已发起 QR 流程，QR 通过事件 `qr.code_generated` 推送
          content:
            application/json:
              schema:
                type: object
                properties:
                  accountId: { type: string }
                  qrSessionId: { type: string }
...
  /v1/admin/workers:
    get:
      tags: [admin]
      summary: 列 worker 集群
      operationId: adminListWorkers
      parameters:
        - in: query
          name: includeDead
          schema: { type: boolean, default: false }
          description: 是否包含心跳超时的 worker（默认仅活着的）
      responses:
        '200':
          description: worker 列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  total: { type: integer }
                  currentWorkerId: { type: string }
                  workers:
                    type: array
                    items:
                      type: object
                      properties:
                        workerId: { type: string }
                        nodeId: { type: string }
                        region: { type: string }
                        capacity: { type: integer }
                        currentLoad: { type: integer }
                        endpoint: { type: string }
                        registeredAt: { type: integer }
  /v1/admin/dead-workers:
    get:
      tags: [admin]
      summary: 列出心跳超时的 worker
      operationId: adminListDeadWorkers
      parameters:
        - in: query
          name: thresholdMs
          schema: { type: integer, default: 60000 }
      responses:
        '200':
          description: dead worker ID 列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  thresholdMs: { type: integer }
                  deadWorkerIds:
                    type: array
                    items: { type: string }
                  checkedAt: { type: string, format: date-time }
...
  /v1/admin/unassign:
    post:
      tags: [admin]
      summary: 手动解除账号 Registry 绑定（兜底）
      operationId: adminUnassign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [accountId]
              properties:
                accountId: { type: string }
                releaseSlot:
                  type: boolean
                  default: true
                  description: |
                    是否同时扣减 Registry load。业务侧收到 account.need_reauth 后确认彻底放弃账号时保持 true；
                    只有确认该账号的 Registry load 已经被其它流程扣过时才传 false。
      responses:
        '200':
          description: 已解除
          content:
            application/json:
              schema:
                type: object
                properties:
                  accountId: { type: string }
                  unassigned: { type: boolean }
                  releaseSlot: { type: boolean }
components:
...
  schemas:
...
    ProxyBinding:
      type: object
      required: [protocol, url, sessionId, country]
      properties:
        protocol:
          type: string
          enum: [socks5, http]
        url:
          type: string
          description: 完整代理 URL（含 session 参数），例如 `socks5://brd-customer-XXX-zone-residential-session-acc_001-country-us:PASS@proxy.example.com:22225`
        sessionId:
          type: string
          description: 持久绑定账号一辈子的 session ID
        provider:
          type: string
          enum: [bright_data, smartproxy, iproyal, custom]
        country: { type: string, example: "US" }
        region: { type: string, nullable: true }
        asn: { type: string, nullable: true, example: "AS7922" }
        tier:
          type: string
          enum: [pairing, standard]
        stickyDurationSec: { type: integer, example: 600 }
...
    DeviceProfile:
      type: object
      description: |
        账号设备元数据。用于功能层筛选、统计、导出和联调展示。
        关联设备展示名请使用 BrowserDisplay。
      required: [accountId, platform, source, deviceCompanion, updatedAt]
      properties:
        accountId: { type: string }
        platform:
          type: string
          enum: [windows, macos, linux, unknown]
          description: 功能层选择或导入材料推断出的桌面设备平台
        source:
          type: string
          enum: [import_body, creds_platform, params_fields, six_fields, legacy_json, unknown]
          description: 设备信息来源
        manufacturer: { type: string, nullable: true, example: Apple }
        model: { type: string, nullable: true, example: MacBook Pro }
        osVersion: { type: string, nullable: true, example: "14.4.1" }
        device: { type: string, nullable: true, example: MacBookPro18,3 }
        deviceUUID: { type: string, nullable: true }
        phoneUUID: { type: string, nullable: true }
        whatsappVersion: { type: string, nullable: true, example: "2.24.10.79" }
        wsDeviceId:
          type: integer
          nullable: true
          description: sixLogin 分身设备 ID；主设备通常为 0
        deviceCompanion:
          type: boolean
          description: 是否分身/伴随设备元数据
        note:
          type: string
          nullable: true
          example: metadata only; does not change Baileys socket browser fingerprint
        updatedAt: { type: string, format: date-time }
    BrowserDisplay:
      type: object
      description: |
        WhatsApp 手机端"已关联设备"里的浏览器展示名。
        协议层会映射为 Baileys browser tuple，例如 `['iOS', 'Opera', '17.5']`，
        手机端可能显示为 `Opera (iOS)`。这只是 Web companion 展示信息，
        不代表协议层运行了真实 Opera 内核或原生 iOS/Android 协议。
      required: [browserName, platform]
      properties:
        accountId: { type: string }
        browserName:
          type: string
          example: Opera
          description: 浏览器展示名，如 Chrome / Opera / Edge / Safari / Firefox
        platform:
          type: string
          enum: [windows, macos, linux, ios, android, unknown]
          example: ios
          description: 关联设备展示平台
        version:
          type: string
          nullable: true
          example: "17.5"
        updatedAt: { type: string, format: date-time }

    # ──── Apifox 三类登录 ────
...